[Feat] 정적 스테이지용 아이템 스폰 매니저 추가

+++ Assets/2.Scripts/3.Object/Item.cs
@@ -5,16 +5,18 @@ using UnityEngine;
 public class Item : MonoBehaviour
 {
     public ItemData data;
-    public int holdingTime = 5;
+    public float holdingTime = 5f;
 
-    private void OnCollisionEnter(Collision other) {
+    private void OnCollisionEnter(Collision other)
+    {
 
         //아이템과 충돌한 플레이어 알려주기
-        if(other.gameObject.tag == "Player"){
+        if (other.gameObject.tag == "Player")
+        {
             var contactedPlayer = other.gameObject.GetComponent<Character>();
             Debug.Log(contactedPlayer + "랑 충돌!");
             //플레이어와 충돌한 아이템 삭제
-           // contactedPlayer.AddItem(itemData)
+            // contactedPlayer.AddItem(itemData)
             //Destroy(gameObject);
         }
     }
@@ -23,7 +25,8 @@ public class Item : MonoBehaviour
 
 
 
-    
+
+
